Separate out landing and about views files
